Methods renamed, commented code removed
