+++ .idea/libraries/junit_junit_4_13.xml
@@ -1,0 +1,11 @@
+<component name="libraryTable">
+  <library name="junit:junit:4.13" type="repository">
+    <properties maven-id="junit:junit:4.13" />
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/junit/junit/4.13/junit-4.13.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
