+++ src/main/task1/task1/Task1.java
@@ -1,4 +1,4 @@
-package task1.task1;
+package main.task1.task1;
 
 import java.io.BufferedReader;
 import java.io.IOException;
