+++ src/main/task1/task1Subtask/PrimeNumberSearch.java
@@ -1,8 +1,11 @@
-package task1.task1;
+package main.task1.task1Subtask;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class PrimeNumberSearch extends Thread {
     int count;
-
+    private static List<Integer> listOfPrimeNumbers = new ArrayList<>();
     private boolean stepCheck = false;
 
     PrimeNumberSearch(int count){
@@ -11,14 +14,18 @@ public class PrimeNumberSearch extends Thread {
 
     public void run(){
         cycle();
+        generateValueParticularlist();
 
+    }
 
+    private void generateValueParticularlist(){
+        Task1.listOfPrimeNumbersParticular.addAll(listOfPrimeNumbers);
     }
 
     private void cycle() {
         for (int i = firstIndex() ; i <= lastIndex(); i++) {
             if (ifPrimeNumber(i) && !isInterrupted()) {
-                Task1.listOfPrimeNumbersGeneral.add(i);
+                listOfPrimeNumbers.add(i);
             }
         }
     }
