+++ src/main/task1/task2/PrimeNumberSearch2.java
@@ -1,4 +1,4 @@
-package task1.task2;
+package main.task1.task2;
 
 import java.util.ArrayList;
 import java.util.List;
