+++ src/main/task2annotation/annotation/StartAnnotationProject.java
@@ -1,4 +1,4 @@
-package task2annotation.annotation;
+package main.task2annotation.annotation;
 
 import java.io.IOException;
