+++ src/main/task2annotation/resourses/annotation.properties
