+++ src/test/task1/task1/TestTask1PrimeNumberSearch.java
@@ -1,0 +1,13 @@
+package test.task1.task1;
+
+import main.task1.task1.PrimeNumberSearch;
+import org.junit.Assert;
+import org.junit.Test;
+
+public class TestTask1PrimeNumberSearch {
+
+
+
+
+
+}
